Add a simple, dumb caching mechanism
Doesn't address the issue of updating an existing post; not without
editing the `.prerender` file manually - which is a gross hack. Should
probably add a "no-cache" option to argparse.

Actually, a "no-cache" option on a per-file basis could circumvent that
particular workflow, at the cost of complecting the post-rendering logic.